# docs/openapi-v2.json
{
  "components": {
    "schemas": {
...
      "Credentials": {
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "setup_token": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "apigate_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HttpOnly browser sessions and CSRF; gateway API keys. Exact decimal USD money, integer usage units.",
    "title": "better-apigate prepaid API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/Credentials"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "attributes": {
                          "properties": {
                            "user_id": {
                              "type": "string"
                            },
                            "verification_required": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "user_id"
                          ],
                          "type": "object"
                        },
                        "id": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "id",
                        "attributes"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [],
        "x-admin-required": false
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": false
      }
    },
...
    "/api/v1/auth/setup": {
      "post": {
        "operationId": "post_auth_setup",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/Credentials"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "attributes": {
                          "properties": {
                            "user_id": {
                              "type": "string"
                            },
                            "verification_required": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "user_id"
                          ],
                          "type": "object"
                        },
                        "id": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "id",
                        "attributes"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [],
        "x-admin-required": false
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}